+++ src/index.ts
@@ -1,129 +1,8 @@
-import stringify from 'json-stable-stringify'
-import { v4 as uuid } from 'uuid'
+export { detectFormat, SERIALIZATION_FORMAT } from './format'
+export { convertDocument, decodeDocument, encodeDocument } from './document'
+export { convertLogic, decodeLogic, encodeLogic } from './logic'
+export { extractProgram, extractProgramFromAST } from './program'
+export { printNode as printMdxNode } from './mdx'
 
-import * as mdx from './mdx'
-
-import { SERIALIZATION_FORMAT, normalizeFormat, detectFormat } from './format'
-import { convertLogic, decodeLogic, encodeLogic } from './serialization'
-
-import * as MDXAST from './types/lona-ast'
-import * as LogicAST from './types/logic-ast'
-
-import { rng } from './utils'
-
-// Document
-
-function decodeDocument(
-  contents: string,
-  format?: SERIALIZATION_FORMAT,
-  filePath?: string
-): { children: MDXAST.Content[] } {
-  const sourceFormat = normalizeFormat(contents, format)
-  try {
-    switch (sourceFormat) {
-      case SERIALIZATION_FORMAT.JSON:
-        return JSON.parse(contents)
-      case SERIALIZATION_FORMAT.SOURCE:
-        return mdx.parse(contents, filePath)
-      default:
-        throw new Error(`Unknown decoding format ${sourceFormat}`)
-    }
-  } catch (e) {
-    console.error(e)
-    throw new Error(`Failed to decode document as ${sourceFormat}.\n\n${e}`)
-  }
-}
-
-function encodeDocument(
-  ast: { children: MDXAST.Content[] },
-  format: SERIALIZATION_FORMAT,
-  options: {
-    sourceFormat?: SERIALIZATION_FORMAT
-    embeddedFormat?: SERIALIZATION_FORMAT
-  } = {}
-) {
-  try {
-    switch (format) {
-      case SERIALIZATION_FORMAT.JSON:
-        return stringify(ast, { space: '  ' })
-      case SERIALIZATION_FORMAT.SOURCE:
-        return mdx.print(ast, options)
-      default:
-        throw new Error(`Unknown encoding format ${format}`)
-    }
-  } catch (e) {
-    console.error(e)
-    throw new Error(`Failed to encode document as ${format}.\n\n${e}`)
-  }
-}
-
-function convertDocument(
-  contents: string,
-  targetFormat: SERIALIZATION_FORMAT,
-  options: {
-    sourceFormat?: SERIALIZATION_FORMAT
-    embeddedFormat?: SERIALIZATION_FORMAT
-    filePath?: string
-  } = {}
-) {
-  const sourceFormat = normalizeFormat(contents, options.sourceFormat)
-
-  if (sourceFormat === targetFormat && !options.embeddedFormat) return contents
-
-  const ast = decodeDocument(contents, sourceFormat, options.filePath)
-  return encodeDocument(ast, targetFormat, options)
-}
-
-function extractProgramFromAST(ast: { children: MDXAST.Content[] }) {
-  const { children } = ast
-
-  const declarations = children
-    .filter(MDXAST.isLonaTokens)
-    // Get Logic syntax node
-    .map(child => child.data.parsed)
-    // Get declarations
-    .map(node => node.data.declarations)
-
-  const flattened = declarations.reduce((prev, x) => prev.concat(x), [])
-
-  const topLevelDeclarations: LogicAST.TopLevelDeclarations = {
-    data: {
-      declarations: flattened,
-      id: uuid({ rng }).toUpperCase(),
-    },
-    type: 'topLevelDeclarations',
-  }
-
-  return topLevelDeclarations
-}
-
-function extractProgram(
-  contents: string,
-  options: { sourceFormat?: SERIALIZATION_FORMAT } = {}
-) {
-  const sourceFormat = normalizeFormat(contents, options.sourceFormat)
-
-  const ast = decodeDocument(contents, sourceFormat)
-
-  const program = extractProgramFromAST(ast)
-
-  return stringify(program, { space: '  ' })
-}
-
-const printMdxNode = mdx.printNode
-
-export {
-  MDXAST,
-  LogicAST,
-  SERIALIZATION_FORMAT,
-  convertLogic,
-  convertDocument,
-  decodeLogic,
-  decodeDocument,
-  encodeLogic,
-  encodeDocument,
-  extractProgram,
-  detectFormat,
-  printMdxNode,
-  extractProgramFromAST,
-}
+export * as LogicAST from './types/logic-ast'
+export * as MDXAST from './types/lona-ast'
